+++ projects/ship/src/main/java/org/openmuc/jeebus/ship/api/Ship.java
@@ -52,6 +52,7 @@ public class Ship implements ShipInterface, AutoCloseable {
      */
     public Ship(ShipConfig nodeConfig, ConnectionHandler connHandler) {
         node = new ShipNodeImpl(nodeConfig, connHandler);
+        node.setClientConnectedListener(this::runConnectionDataPreparation);
     }
 
     /**